KVM: PPC: Book3S HV: Nested: Implement nest rmap invalidations for hpt guests
The nest rmap is used to store a reverse mapping from the (L1) guest real
address back to a pte in the shadow page table which maps it. This is used
when the host is modifying a L1 guest pte (either invalidating it or
modifying the rc bits) to make the necessary changes to the ptes in
the shadow tables which map that L1 guest page. This is already
implemented for a nested radix guest where the rmap entry stores the gpa
(guest physical address) of the nested pte which can be used to traverse
the shadow page table and find any matching ptes. Implement this nested
rmap invalidation for nested hpt (hash page table) guests.

We reuse the nest rmap structure that already exists for radix nested
guests for nested hpt guests. Instead of storing the gpa the hpt index
of the pte is stored. This means that a pte in the shadow hpt can be
uniquely identified by the nest rmap. As with the radix case we check
that the same host page is being addressed to detect if this is a stale
rmap entry, in which case we skip the invalidation.

When the host is invalidating a mapping for a L1 guest page use the
nest rmap to find any shadow ptes in the shadow hpt which map that page
and make them absent, and invalidate any caching of the entry. By making
the pte absent rather than zeroing it we can avoid having to lookup the
guest rpte the next time an entry is faulted in.

When the host is clearing rc bits for a mapping for a L1 guest page use
the nest rmap to find any shadow ptes in the shadow hpt which map that
page and make them absent and invalidate them as in the above case for
invalidating a L1 guest page. It is not sufficient to clear the rc bits
in the shadow pte since hardware can set them again without software
intervention, so the mapping must be made invalid so that we will take
a page fault and can ensure that the rc bits stay in sync in the page
fault handler.

Signed-off-by: Suraj Jitindar Singh <sjitindarsingh@gmail.com>